+++ src/main/java/org/jabref/gui/JabRefFrame.java
@@ -904,8 +904,7 @@ public class JabRefFrame extends BorderPane {
 
                 new SeparatorMenuItem(),
 
-                factory.createMenuItem(StandardActions.SEND_AS_EMAIL, new SendAsEMailAction(StandardActions.SEND_AS_EMAIL, dialogService, this.prefs, stateManager)),
-                factory.createMenuItem(StandardActions.SEND_TO_KINDLE, new SendAsEMailAction(StandardActions.SEND_TO_KINDLE, dialogService, this.prefs, stateManager)),
+                createSendSubMenu(factory, dialogService, stateManager, prefs),
                 pushToApplicationMenuItem,
 
                 new SeparatorMenuItem(),
@@ -1321,6 +1320,19 @@ public class JabRefFrame extends BorderPane {
         destinationLibraryTab.getBibDatabaseContext()
                              .getMetaData()
                              .setGroups(currentLibraryGroupRoot);
+    }
+
+    private Menu createSendSubMenu(ActionFactory factory,
+                                          DialogService dialogService,
+                                          StateManager stateManager,
+                                          PreferencesService preferencesService) {
+        Menu sendMenu = factory.createMenu(StandardActions.SEND);
+        sendMenu.getItems().addAll(
+                factory.createMenuItem(StandardActions.SEND_AS_EMAIL, new SendAsGeneralEmailAction(dialogService, preferencesService, stateManager)),
+                factory.createMenuItem(StandardActions.SEND_TO_KINDLE, new SendAsKindleEmailAction(dialogService, preferencesService, stateManager))
+        );
+
+        return sendMenu;
     }
 
     /**